+++ scripts/ci/50_make_image_fedora.sh
@@ -101,9 +101,11 @@ Language=zh_CN.UTF-8
 SystemAccount=true
 EOF
 
-install -d -m 0755 /home/user/.config
-install -Dm644 /usr/local/share/gaokun/monitors.xml /home/user/.config/monitors.xml
-chown user:user /home/user/.config/monitors.xml
+if [[ -f /usr/local/share/gaokun/monitors.xml ]]; then
+  install -d -m 0755 /home/user/.config
+  install -Dm644 /usr/local/share/gaokun/monitors.xml /home/user/.config/monitors.xml
+  chown user:user /home/user/.config/monitors.xml
+fi
 
 systemctl enable gdm NetworkManager sshd huawei-touchpad.service \
   gdm-monitor-sync.service patch-nvm-bdaddr.service || true
